+++ docs/pages/Buttons.jsx
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 import { Display3 } from '../../src';
-import ButtonsTypes from '../examples/Buttons/Buttons';
+import ButtonsPage from '../examples/Buttons/Buttons';
 
 /**
  * @return {React.Component}
@@ -9,7 +9,7 @@ export const Buttons = () => {
     return (
         <Fragment>
             <Display3>Buttons</Display3>
-            <ButtonsTypes />
+            <ButtonsPage />
         </Fragment>
     );
 };
