+++ package.json
@@ -40,8 +40,10 @@
     "faker": "^4.1.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^24.9.0",
+    "node-sass": "^4.12.0",
     "prismjs": "^1.17.1",
     "react-prism": "^4.3.2",
+    "sass-loader": "^8.0.0",
     "style-loader": "^1.0.0",
     "webpack": "^4.39.3",
     "webpack-cli": "^3.3.8",
