+++ src/Buttons/Buttons.scss
@@ -1,0 +1,168 @@
+.btn {
+    margin: 3px;
+}
+
+.btn-color-default {
+    background-color: transparent;
+    color: #3c3c3c;
+    &:hover {
+        color: #000;
+        background-color: transparent;
+        border: 1px solid #727272;
+    }
+    &:focus {
+        color: #000;
+        background-color: transparent;
+        border: 1px solid #727272;
+    }
+}
+
+.btn-color-red {
+    background-color: #ea0000;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #ac151d;
+        border: 1px solid #ac151d;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #ac151d;
+        border: 1px solid #ac151d;
+    }
+}
+
+.btn-color-green {
+    background-color: green;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #237535;
+        border: 1px solid #237535;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #237535;
+        border: 1px solid #237535;
+    }
+}
+
+.btn-color-blue {
+    background-color: #0042b9;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #003996;
+        border: 1px solid #003996;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #003996;
+        border: 1px solid #003996;
+    }
+}
+
+.btn-color-yellow {
+    background-color: #d4b500;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #bd9e00;
+        border: 1px solid #bd9e00;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #bd9e00;
+        border: 1px solid #bd9e00;
+    }
+}
+
+.btn-color-purple {
+    background-color: #ae008f;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #86006c;
+        border: 1px solid #86006c;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #86006c;
+        border: 1px solid #86006c;
+    }
+}
+
+.btn-color-pink {
+    background-color: #e56dba;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #be5b9a;
+        border: 1px solid #be5b9a;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #be5b9a;
+        border: 1px solid #be5b9a;
+    }
+}
+
+.btn-color-deeppink {
+    background-color: #e820a1;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #ba1980;
+        border: 1px solid #ba1980;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #ba1980;
+        border: 1px solid #ba1980;
+    }
+}
+
+.btn-color-orange {
+    background-color: #e97900;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #c26500;
+        border: 1px solid #c26500;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #c26500;
+        border: 1px solid #c26500;
+    }
+}
+
+.btn-color-grey {
+    background-color: #888181;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #696464;
+        border: 1px solid #696464;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #696464;
+        border: 1px solid #696464;
+    }
+}
+
+.btn-color-black {
+    background-color: #111111;
+    color: #fff;
+    &:hover {
+        color: #fff;
+        background-color: #080808;
+        border: 1px solid #080808;
+    }
+    &:focus {
+        color: #fff;
+        background-color: #080808;
+        border: 1px solid #080808;
+    }
+}
